Set all DateTime.Parse methods to use CurrentCulture.Invariant to support globalization.  Also, updated type for status IDs to be ulong.  I already did Twypocalypse updates using uint, but was convinced by Twitter API developer team that 64-bit unsigned (ulong in C#) is a much better choice.

+++ LinqToTwitter/Blocks.cs
@@ -18,12 +18,12 @@ namespace LinqToTwitter
         /// <summary>
         /// id of user
         /// </summary>
-        public uint ID { get; set; }
+        public ulong ID { get; set; }
 
         /// <summary>
         /// disambiguates when user id is screen name
         /// </summary>
-        public uint UserID { get; set; }
+        public ulong UserID { get; set; }
 
         /// <summary>
         /// disambiguates when screen name is user id
